--- src/main/java/com/example/community_cr/comment_like/repository/CommentLikeRepository.java
@@ -1,10 +1,0 @@
-package com.example.community_cr.comment_like.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.example.community_cr.comment_like.entity.CommentHeart;
-import com.example.community_cr.comment_like.entity.CommentHeartId;
-
-public interface CommentLikeRepository extends JpaRepository<CommentHeart, CommentHeartId> {
-	boolean existsByUserIdAndCommentId(Long userId, Long commentId);
-}
