--- src/main/java/com/example/community_cr/comment_like/service/CommentLikeService.java
@@ -1,7 +1,0 @@
-package com.example.community_cr.comment_like.service;
-
-public interface CommentLikeService {
-    void likeComment(long userId, long commentId);
-    void unlikeComment(long userId, long commentId);
-}
-
