--- src/main/java/com/example/community_cr/community/block/repository/BlockRepository.java
@@ -1,9 +1,0 @@
-package com.example.community_cr.community.block.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.example.community_cr.community.block.entity.Block;
-import com.example.community_cr.community.block.entity.BlockId;
-
-public interface BlockRepository extends JpaRepository<Block, BlockId> {
-}
