--- src/main/java/com/example/community_cr/community/block/service/BlockService.java
@@ -1,11 +1,0 @@
-package com.example.community_cr.community.block.service;
-
-import java.util.Optional;
-
-import com.example.community_cr.community.post.controller.dto.response.PostResponse;
-
-public interface BlockService {
-	void blockPost(long userId, long postId);
-
-	Optional<PostResponse> unblockPost(long userId, long postId);
-}
