--- src/main/java/com/example/community_cr/community/like/service/LikeService.java
@@ -1,7 +1,0 @@
-package com.example.community_cr.community.like.service;
-
-public interface LikeService {
-	void likePost(long senderId, long postId);
-
-	void unlikePost(long senderId, long postId);
-}
